verificar terminado

+++ ProyectoNet/General/Csv/ReciboFirmado.cs
@@ -8,7 +8,7 @@ namespace General.Csv
     public class CSV
     {
 
-        public int id { get; set; }
+        public long id { get; set; }
         //public DateTime fechaConformidadUsuario { get; set; }
         public string csv { get; set; }
         public int tipo_doc_electronico { get; set; }
@@ -17,13 +17,13 @@ namespace General.Csv
         public CSV()
         {
         }
-        public CSV(int id, string csv, int tipo_doc_electronico, int id_doc)
+        public CSV(long id, string csv, int tipo_doc_electronico, int id_doc)
         {
             this.id = id;
             this.csv = csv;
             this.tipo_doc_electronico = tipo_doc_electronico;
             this.id_doc = id_doc;
-
+            
         }
 
     }
